102-fibonacci.c

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,7 +7,7 @@
 int main(void) {
     int n = 50;
     int i;
-    unsigned long long int fib[n];
+    unsigned int fib[n];
 
     fib[0] = 1;
     fib[1] = 2;
